[Peer Monitoring Service] Refactor request/response messages.
