+++ network/peer-monitoring-service/client/src/error.rs
@@ -2,7 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use aptos_network::protocols::network::RpcError;
-use aptos_peer_monitoring_service_types::{PeerMonitoringServiceError, UnexpectedResponseError};
+use aptos_peer_monitoring_service_types::{
+    response::UnexpectedResponseError, PeerMonitoringServiceError,
+};
 use thiserror::Error;
 
 #[derive(Debug, Error)]
