+++ network/peer-monitoring-service/client/src/peer_states/key_value.rs
@@ -12,7 +12,8 @@ use aptos_config::{config::NodeConfig, network_id::PeerNetworkId};
 use aptos_infallible::RwLock;
 use aptos_network::application::metadata::PeerMetadata;
 use aptos_peer_monitoring_service_types::{
-    LatencyPingRequest, PeerMonitoringServiceRequest, PeerMonitoringServiceResponse,
+    request::{LatencyPingRequest, PeerMonitoringServiceRequest},
+    response::PeerMonitoringServiceResponse,
 };
 use aptos_time_service::TimeService;
 use enum_dispatch::enum_dispatch;
