+++ network/peer-monitoring-service/types/Cargo.toml
@@ -14,6 +14,6 @@ rust-version = { workspace = true }
 
 [dependencies]
 aptos-config = { workspace = true }
-aptos-network = { workspace = true }
+aptos-types = { workspace = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
