+++ network/src/application/metadata.rs
@@ -5,9 +5,8 @@ use crate::{
     protocols::wire::handshake::v1::{ProtocolId, ProtocolIdSet},
     transport::ConnectionMetadata,
 };
-use aptos_config::network_id::PeerNetworkId;
+use aptos_peer_monitoring_service_types::PeerMonitoringMetadata;
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 
 /// The current connection state of a peer
 /// TODO: Allow nodes that are unhealthy to stay connected
@@ -16,32 +15,6 @@ pub enum ConnectionState {
     Connected,
     Disconnecting,
     Disconnected, // Currently unused (TODO: fix this!)
-}
-
-/// The peer monitoring metadata for a peer
-#[derive(Clone, Debug, Default, Deserialize, PartialEq, Serialize)]
-pub struct PeerMonitoringMetadata {
-    pub average_ping_latency_secs: Option<f64>, // The average latency ping for the peer
-    pub connected_peers: Option<HashMap<PeerNetworkId, ConnectionMetadata>>, // Connected peers and metadata
-    pub distance_from_validators: Option<u64>, // The known distance from the validator set
-}
-
-/// We must manually define this because f64 doesn't implement Eq. Instead,
-/// we rely on PartialEq (which is sufficient for our use-cases).
-impl Eq for PeerMonitoringMetadata {}
-
-impl PeerMonitoringMetadata {
-    pub fn new(
-        average_ping_latency_secs: Option<f64>,
-        connected_peers: Option<HashMap<PeerNetworkId, ConnectionMetadata>>,
-        distance_from_validators: Option<u64>,
-    ) -> Self {
-        PeerMonitoringMetadata {
-            average_ping_latency_secs,
-            connected_peers,
-            distance_from_validators,
-        }
-    }
 }
 
 /// A container holding all relevant metadata for the peer.
